+++ services/opentsdb/service.go
@@ -74,7 +74,7 @@ type Service struct {
 	Logger         *log.Logger
 
 	stats    *Statistics
-	statTags models.Tags
+	statTags map[string]string
 }
 
 // NewService returns a new instance of Service.
@@ -96,7 +96,7 @@ func NewService(c Config) (*Service, error) {
 		Logger:          log.New(os.Stderr, "[opentsdb] ", log.LstdFlags),
 		LogPointErrors:  d.LogPointErrors,
 		stats:           &Statistics{},
-		statTags:        models.NewTags(map[string]string{"bind": d.BindAddress}),
+		statTags:        map[string]string{"bind": d.BindAddress},
 	}
 	return s, nil
 }
@@ -200,9 +200,16 @@ type Statistics struct {
 
 // Statistics returns statistics for periodic monitoring.
 func (s *Service) Statistics(tags map[string]string) []models.Statistic {
+	// Insert any missing deault tag values.
+	for k, v := range s.statTags {
+		if _, ok := tags[k]; !ok {
+			tags[k] = v
+		}
+	}
+
 	return []models.Statistic{{
 		Name: "opentsdb",
-		Tags: s.statTags,
+		Tags: tags,
 		Values: map[string]interface{}{
 			statHTTPConnectionsHandled:   atomic.LoadInt64(&s.stats.HTTPConnectionsHandled),
 			statTelnetConnectionsActive:  atomic.LoadInt64(&s.stats.ActiveTelnetConnections),